Added NOP functionality
Added a few functions and tests. Currently starts by adding the NOP command at address 7000. Then prompts user for the start address. User should enter 7000 to get the test to work. Then the program will decode the operation at address 7000 and if it is NOP, it will print NOP to the screen.